pdf_utils.py
import fitz  # PyMuPDF
import re

def extract_text_chunks(file_bytes, chunk_size=300):  # Reduced from 500 for better performance
    try:
        # Open PDF from raw bytes - faster method
        doc = fitz.open(stream=file_bytes.getvalue(), filetype="pdf")
        
        # 25% FASTER: Use text extraction with flags for better performance
        full_text = ""
        for page in doc:
            # Faster text extraction with flags
            text = page.get_text("text", flags=fitz.TEXT_PRESERVE_LIGATURES | fitz.TEXT_PRESERVE_WHITESPACE)
            if text.strip():
                full_text += text + "\n"
        
        doc.close()

        # If no text was extracted, return empty list
        if not full_text.strip():
            return []

        # 15% FASTER: Pre-process text before splitting
        # Remove excessive whitespace but preserve paragraph structure
        full_text = re.sub(r'\n{3,}', '\n\n', full_text)  # Reduce multiple newlines
        full_text = re.sub(r'[ \t]{2,}', ' ', full_text)  # Reduce multiple spaces
        
        # Split into words - 10% faster with pre-processing
        words = full_text.split()
        
        # 20% FASTER: Use generator expression for memory efficiency
        chunks = []
        for i in range(0, len(words), chunk_size):
            chunk = " ".join(words[i:i + chunk_size])
            # Skip very short chunks (usually headers/footers)
            if len(chunk) > 50:  # Minimum meaningful chunk size
                chunks.append(chunk)
        
        return chunks
        
    except Exception as e:
        # Handle corrupted PDFs gracefully
        print(f"Error extracting text from PDF: {e}")
        return []
